requirements.txt
asgiref==3.6.0
Django==4.1.7
django-ninja==0.21.0
pydantic==1.10.7
sqlparse==0.4.3
typing_extensions==4.5.0
